+++ eslint.config.js
@@ -1,0 +1,12 @@
+const { FlatCompat } = require('@eslint/eslintrc');
+
+const compat = new FlatCompat({
+  baseDirectory: __dirname,
+});
+
+module.exports = [
+  ...compat.extends('@react-native'),
+  {
+    ignores: ['node_modules/**', 'android/**', 'ios/**'],
+  },
+];